microservices/auth-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
redis==5.0.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0